// ts/bootstrap.mobile-nav.d.ts
export interface IMobileNavOptions {
    collapse: string;
    backdropIn: string;
    backdropOut: string;
    closeId: string;
    closeBtn: string;
    minWidthHidden: number;
    resizeDelay: number;
    hideOnClick: boolean;
    onClickLink: string;
}

export interface IMobileNavDefaultSettings {
    settings: IMobileNavOptions;
}